fix(host): constrain UDP packetization to 1200B and resolve Windows DXGI physical mode
- Constrain UDP video chunk size to 1,154 B and audio fragment size to 1,152 B,
  limiting datagram payload to 1,200 B to eliminate IP fragmentation across 1,280 B MTU tunnels.
- Derive physical display dimensions from DXGI_OUTDUPL_DESC.ModeDesc while preserving
  logical DesktopCoordinates for client hit-testing, preventing NV12 length mismatch crashes on high-DPI Windows hosts.
- Add bounded capture-to-encode stage attribution instrumentation.
- Add unit regression tests in sender_packetization_tests.rs and sender_trace_tests.rs.

+++ clients/rust/erd-host/src/audio_windows.rs
@@ -116,7 +116,7 @@ mod imp {
                     break;
                 }
                 let frame_bytes =
-                    self.source.channels as usize * if self.source.is_float { 4 } else { 2 };
+                    self.source.channels * if self.source.is_float { 4 } else { 2 };
                 if flags & AUDCLNT_BUFFERFLAGS_SILENT != 0 {
                     // Buffer contents are undefined while the flag is set.
                     raw.resize(raw.len() + frames as usize * frame_bytes, 0);
